[Flash] Fix availability of flash pins on ESP32C3
